+++ CHANGES.md
@@ -33,6 +33,7 @@ Release Notes.
 - Add node labels to classify the nodes.
 - Add lifecycle management for the node.
 - Property: Introduce the schema style to the property.
+- Add time range parameters to stream index filter.
 
 ### Bug Fixes